requirements.txt
python==3.11
pyautogui
pynput
click
pillow